requirements.txt
# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# ElevenLabs SDK
elevenlabs>=0.2.27

# Audio Processing
sounddevice>=0.4.6
soundfile>=0.12.1
# pydub>=0.25.1  # Removed - not compatible with Python 3.13
numpy>=1.24.0
scipy>=1.11.0

# Translation
deep-translator>=1.11.4
langdetect>=1.0.9

# HTTP and API
requests>=2.31.0
aiohttp>=3.9.0

# File Management
pathlib>=1.0.1

# UI Components
# streamlit-audio-recorder>=0.0.8  # Optional, not available in PyPI
streamlit-option-menu>=0.3.6

# Utilities
python-multipart>=0.0.6
typing-extensions>=4.8.0

# Development Tools (optional)
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
pylint>=2.17.0
